calculate md5 for parallel uploads
