+++ src/Blob/Helpers/DateHelper.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace AzureOss\Storage\Blob\Helpers;
 
+use AzureOss\Storage\Blob\Exceptions\DeserializationException;
+
 /**
  * @internal
  */
@@ -14,5 +16,15 @@ final class DateHelper
         return \DateTime::createFromInterface($date)
             ->setTimezone(new \DateTimeZone('UTC'))
             ->format('Y-m-d\TH:i:s\Z');
+    }
+
+    public static function deserializeDateRfc1123Date(string $date): \DateTimeInterface
+    {
+        $result = \DateTimeImmutable::createFromFormat(\DateTimeInterface::RFC1123, $date);
+        if ($result === false) {
+            throw new DeserializationException("Azure returned a malformed date.");
+        }
+
+        return $result;
     }
 }
